pyproject.toml
[project]
name = "multiagent-comparison"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastparquet>=2026.3.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.2",
    "langchain-google-genai>=4.2.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.1",
    "matplotlib>=3.10.9",
    "mkl-service>=2.6.1",
    "pandas>=3.0.2",
    "pyarrow>=24.0.0",
    "python-dotenv>=1.2.2",
    "tabulate>=0.9.0",
    "tqdm>=4.67.3",
]
